Set values during recording pauses to zero (important update)
HANDLE RECORDING PAUSES IN EYE-TRACKING DATA

If the ET recording was paused (generally not recommended) there will be jumps in the timestamp. During the linear interpolation (see below) the ET signal during these gaps will be also be linearly interpolated  (i.e. there will be a linear trend between the last sample before and first sample after the pause. This is not helpful, because these
intervals will not be easily detected as missing data (e.g. by saccade  detection algorithm or by rej_eyecontin.m), because they contain plausible non-zero values.

This next step therefore detects jumps in the ET time stamp and sets the two samples adjacent to the recording pause to zero, so the pause will just contain zeros, event after interpolation and will be easily.

Note: the threshold vor a "pause" was arbitraily set to the duration of three median inter-sample intervals. If your eye-tracker produces larger irregularites in the time stamps even without a pause, this values would need to be set higher.